+++ src/Component/Form/UITextareaComponent.php
@@ -79,7 +79,7 @@ final class UITextareaComponent extends AbstractComponent
             'field' => $this->name ?: $this->id,
             'value' => $this->value,
             'error' => $this->errorMessage,
-            'isValid' => empty($this->errorMessage)
+            'isValid' => (int) empty($this->errorMessage)
         ]);
     }
